Fixes some warnings throughout and fixes an issue where heat flux was not being re-dimensionalized for history file output.

+++ Common/src/fem_gauss_jacobi_quadrature.cpp
@@ -88,7 +88,7 @@ void CGaussJacobiQuadrature::GetQuadraturePoints(const passivedouble   alpha,   
                                                  vector<passivedouble> &GJPoints, vector<passivedouble> &GJWeights) {
 
   /*--- Determine the number of integration points. Check if the number makes sense. ---*/
-  unsigned int nIntPoints = GJPoints.size();
+  unsigned int nIntPoints = (unsigned int)GJPoints.size();
   if(nIntPoints < 1 || nIntPoints > 100)
     SU2_MPI::Error("Invalid number of Gauss Jacobi integration points", CURRENT_FUNCTION);
